+++ src/components/Header.tsx
@@ -1,4 +1,4 @@
-import UserProfile from "@/components/UserProfile";
+import { UserProfile } from "@/components/UserProfile";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { useState } from "react";
@@ -6,7 +6,7 @@ import { GoLinkExternal, GoMarkGithub } from "react-icons/go";
 import { IconType } from "react-icons/lib";
 import { TbHome, TbInfoSquareRounded } from "react-icons/tb";
 
-export default function Header() {
+export function Header() {
     return (
         <header className="flex items-center justify-end gap-4 bg-black-light px-10 py-3 text-lg text-gray-lightest">
             <h1 className="mr-auto text-2xl font-extrabold text-primary">Stupid Spotify</h1>
