+++ src/components/Player/index.ts
