--- src/pages/test.md
@@ -1,3 +1,0 @@
-# testing
-
-Testing tests
